refactor code to send required data to client side Relates #11
