+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "node src/test/test.js | tap-spec",
-    "dev": "nodemon src/index.js" 
+    "dev": "nodemon src/index.js"
   },
   "repository": {
     "type": "git",
@@ -23,7 +23,8 @@
     "env2": "^2.2.2",
     "express": "^4.17.1",
     "express-handlebars": "^3.1.0",
-    "node-fetch": "^2.6.0"
+    "node-fetch": "^2.6.0",
+    "request": "^2.88.0"
   },
   "devDependencies": {
     "eslint": "^5.16.0",
