+++ public/dom.js
@@ -1,18 +1,20 @@
 const search = document.querySelector('#search');
 const userinput = document.querySelector('#userInput');
-const url = '/search';
+
 search.addEventListener('submit', (e) => {
 
   e.preventDefault();
   const input = userinput.value.trim();
+  console.log(input)
   if (!input) {
     return;
   }
-  fetch('/search', {
+   fetch('/search', {
       method:"POST",
       headers : {'Content-Type':'application/json'},
       body:JSON.stringify({input})
   })
     .then(res => res.json())
-    .then(data=> console.log(data));
+    .then(data=> console.log(data))      
 });
+
