+++ src/app.js
@@ -1,10 +1,11 @@
 const path = require('path');
 const express = require('express');
 const exhbs = require('express-handlebars');
-const helpers = require('./controlers/index');
 const router = require('./controlers/index');
 const app = express();
+
 require('env2')('.env')
+app.disable('x-powered-by');
 
 app.use(express.static(path.join(__dirname, '..', 'public')));
 app.use(express.json());
@@ -18,8 +19,7 @@ app.engine(
         extname:'hbs',
         layoutsDir:path.join(__dirname,'views','layouts'),
         partialsDir: path.join(__dirname, 'views', 'partials'),
-        defaultLayout: 'main',
-        helpers:helpers
+        defaultLayout: 'main'
     })
 )
