+++ src/test/test.js
@@ -1,7 +1,18 @@
 
 const test = require('tape');
+const supertest = require('supertest');
+const app = require('../app');
 
-test("testing",t=> {
-    t.equal(5,5,'should return 5');
-    t.end();
+
+test("testing for status code 404",t=>{
+    supertest(app)
+    .get('/test')
+    .expect(404)
+    .expect("Content-Type", /html/)
+    .end((err,res)=> {
+        t.error(err)
+        t.equal(res.text,'Page Not Found','should return not found')
+        t.end()
+    })
 })
+
