+++ src/views/home.hbs
@@ -1,7 +1,15 @@
 
-<h1>{{appName}}</h1>
-<form id="search" >
-    <input id="userInput" type="text" value="" name = "input" placeholder= "Search for your favourite cat" required>
-    <input type="submit">
-</form>
+<header>
+    <h1>{{appName}}</h1>
+</header>
 
+<section>
+    <form id="search" >
+        <input id="userInput" type="text" value="" name = "input" placeholder= "Search for your favourite cat" required>
+        <input type="submit">
+    </form>
+</section>
+
+<section>
+    <div><img src ='' alt="cat photo"></div>
+</section>
